+++ src/render/render_ui_texture.c
@@ -6,14 +6,14 @@
 /*   By: njennes <njennes@student.42lyon.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/14 12:37:18 by njennes           #+#    #+#             */
-/*   Updated: 2022/05/16 23:36:38 by njennes          ###   ########.fr       */
+/*   Updated: 2022/05/18 15:04:22 by njennes          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "texture.h"
 #include "core.h"
 
-void	render_ui_texture(int64_t tex_id, int x, int y)
+void	render_ui_texture(int64_t tex_id, int64_t x, int64_t y)
 {
 	t_mlx				*app;
 	t_texture			*texture;
@@ -23,10 +23,10 @@ void	render_ui_texture(int64_t tex_id, int x, int y)
 		return ;
 	app = get_app();
 	mlx_put_image_to_window(app->mlx, app->win,
-			texture->original_handle, x, y);
+			texture->original_handle, (int)x, (int)y);
 }
 
-void	render_ui_texture_grayscale(int64_t tex_id, int x, int y)
+void	render_ui_texture_grayscale(int64_t tex_id, int64_t x, int64_t y)
 {
 	t_mlx				*app;
 	t_texture			*texture;
@@ -36,10 +36,10 @@ void	render_ui_texture_grayscale(int64_t tex_id, int x, int y)
 		return ;
 	app = get_app();
 	mlx_put_image_to_window(app->mlx, app->win,
-			texture->grayscale_handle, x, y);
+			texture->grayscale_handle, (int)x, (int)y);
 }
 
-void	render_ui_texture_inverse_grayscale(int64_t tex_id, int x, int y)
+void	render_ui_texture_inverse_grayscale(int64_t tex_id, int64_t x, int64_t y)
 {
 	t_mlx				*app;
 	t_texture			*texture;
@@ -49,10 +49,10 @@ void	render_ui_texture_inverse_grayscale(int64_t tex_id, int x, int y)
 		return ;
 	app = get_app();
 	mlx_put_image_to_window(app->mlx, app->win,
-			texture->inversed_grayscale_handle, x, y);
+			texture->inversed_grayscale_handle, (int)x, (int)y);
 }
 
-void	render_ui_texture_inversed(int64_t tex_id, int x, int y)
+void	render_ui_texture_inversed(int64_t tex_id, int64_t x, int64_t y)
 {
 	t_mlx				*app;
 	t_texture			*texture;
@@ -62,5 +62,5 @@ void	render_ui_texture_inversed(int64_t tex_id, int x, int y)
 		return ;
 	app = get_app();
 	mlx_put_image_to_window(app->mlx, app->win,
-			texture->inversed_handle, x, y);
+			texture->inversed_handle, (int)x, (int)y);
 }
